+++ packages/auto_animated/lib/auto_animated.dart
@@ -2,6 +2,6 @@ export 'src/grid.dart';
 export 'src/helpers/callbacks.dart';
 export 'src/icon_button.dart';
 export 'src/list.dart';
-export 'src/on_visibility_change.dart';
+export 'src/animate_if_visible.dart';
 export 'src/sliver_grid.dart';
 export 'src/sliver_list.dart';